% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_download_and_clean.R
\name{cbi_download}
\alias{cbi_download}
\title{Download Monthly Card Payment Statistics}
\usage{
cbi_download(dest_path)
}
\arguments{
\item{dest_path}{The path where the Excel file should be saved, including the
file name (e.g., "./data/my_payments.xlsx")}
}
\value{
The function downloads the file and returns the file path invisibly.
}
\description{
This function downloads the latest monthly card payment statistics
from the Central Bank of Ireland website. It will only download a new file
if the content has changed since the last download, this is convenient for a
\code{targets} workflow as an example.
}
