% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_download_and_clean.R
\name{cbi_get_dl_link}
\alias{cbi_get_dl_link}
\title{Scrape the Download URL for the CBI Monthly Card Payment Statistics}
\usage{
cbi_get_dl_link()
}
\value{
The latest URL for the CBI Monthly Card Payment Statistics
}
\description{
This function web scrapes the latest download URL for the monthly card
payment statistics .xlsx file from the Central Bank of Ireland website.
}
