% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbi_download_and_clean.R
\name{cbi_read_data}
\alias{cbi_read_data}
\title{Read and Clean Card Payment Statistics}
\usage{
cbi_read_data(file_path)
}
\arguments{
\item{file_path}{The path to the Excel file containing the data.}
}
\value{
A clean \code{tibble} or \code{data.frame} with the payment statistics.
}
\description{
This function reads the "Master Data" sheet from the monthly card payments
Excel file, formats the \verb{Reporting Period} and removes the \code{NA} values from
\code{Sector} and \code{Sub-sector}.
}
