% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate.R
\name{validate_cbi_download}
\alias{validate_cbi_download}
\title{Validate CBI data download}
\usage{
validate_cbi_download(data)
}
\arguments{
\item{data}{A data downloaded from \code{\link[=cbi_download]{cbi_download()}}.}
}
\value{
Invisibly returns \code{data} if valid, otherwise throws an error.
}
\description{
Ensures that the downloaded data has the expected structure
before being passed to \code{\link[=cbi_read_data]{cbi_read_data()}}.
}
\keyword{internal}
