% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate.R
\name{validate_characters}
\alias{validate_characters}
\title{Check if multiple arguments are character strings}
\usage{
validate_characters(...)
}
\arguments{
\item{...}{A series of arguments to be checked.}
}
\value{
The function returns \code{invisible(TRUE)} if all arguments are
character strings, otherwise it throws an error.
}
\description{
Check if multiple arguments are character strings
}
