% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate.R
\name{validate_single_string}
\alias{validate_single_string}
\title{Check if an argument is a single character string}
\usage{
validate_single_string(...)
}
\arguments{
\item{...}{A series of arguments to be checked.}
}
\value{
The function returns \code{invisible(TRUE)} if the argument is a single
character string, otherwise it throws an error.
}
\description{
Check if an argument is a single character string
}
